- more solaris fixes, should now build and install cleanly
git-svn-id: http://svn.ratbox.org/svnroot/trunk@3585 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/match.c
@@ -215,7 +215,7 @@ unsigned long textip_to_ul(const char *ip)
 
   char c;
   while((c=*ip)) {
-    if(isdigit(c)) {
+    if(isdigit((int)c)) {
       octet *= 10;
       octet += (*ip & 0xF);
     } else if(c == '.') {
